combine deeptaubase with deetauid
