+++ RecoTauTag/RecoTau/plugins/DeepTauId.cc
@@ -7,7 +7,28 @@
  *         Christian Veelken, Tallinn
  */
 
-#include "RecoTauTag/RecoTau/interface/DeepTauBase.h"
+#include <Math/VectorUtil.h>
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "PhysicsTools/TensorFlow/interface/TensorFlow.h"
+#include "tensorflow/core/util/memmapped_file_system.h"
+#include "DataFormats/PatCandidates/interface/Electron.h"
+#include "DataFormats/PatCandidates/interface/Muon.h"
+#include "DataFormats/PatCandidates/interface/Tau.h"
+#include "DataFormats/TauReco/interface/TauDiscriminatorContainer.h"
+#include "DataFormats/TauReco/interface/PFTauDiscriminator.h"
+#include "DataFormats/PatCandidates/interface/PATTauDiscriminator.h"
+#include "CommonTools/Utils/interface/StringObjectFunction.h"
+#include "RecoTauTag/RecoTau/interface/PFRecoTauClusterVariables.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+#include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
+#include "DataFormats/Common/interface/View.h"
+#include "DataFormats/Common/interface/RefToBase.h"
+#include "DataFormats/Provenance/interface/ProductProvenance.h"
+#include "DataFormats/Provenance/interface/ProcessHistoryID.h"
+#include "FWCore/Common/interface/Provenance.h"
+#include <TF1.h>
+#include <map>
 #include "RecoTauTag/RecoTau/interface/DeepTauScaling.h"
 #include "FWCore/Utilities/interface/isFinite.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
@@ -17,8 +38,123 @@
 #include "oneapi/tbb/concurrent_unordered_set.h"
 
 namespace deep_tau {
+  enum BasicDiscriminator {
+    ChargedIsoPtSum,
+    NeutralIsoPtSum,
+    NeutralIsoPtSumWeight,
+    FootprintCorrection,
+    PhotonPtSumOutsideSignalCone,
+    PUcorrPtSum
+  };
+
+  class TauWPThreshold {
+  public:
+    explicit TauWPThreshold(const std::string& cut_str) {
+      bool simple_value = false;
+      try {
+        size_t pos = 0;
+        value_ = std::stod(cut_str, &pos);
+        simple_value = (pos == cut_str.size());
+      } catch (std::invalid_argument&) {
+      } catch (std::out_of_range&) {
+      }
+      if (!simple_value) {
+        static const std::string prefix =
+            "[&](double *x, double *p) { const int decayMode = p[0];"
+            "const double pt = p[1]; const double eta = p[2];";
+        static const int n_params = 3;
+        static const auto handler = [](int, Bool_t, const char*, const char*) -> void {};
+
+        std::string fn_str = prefix;
+        if (cut_str.find("return") == std::string::npos)
+          fn_str += " return " + cut_str + ";}";
+        else
+          fn_str += cut_str + "}";
+        auto old_handler = SetErrorHandler(handler);
+        fn_ = std::make_unique<TF1>("fn_", fn_str.c_str(), 0, 1, n_params);
+        SetErrorHandler(old_handler);
+        if (!fn_->IsValid())
+          throw cms::Exception("TauWPThreshold: invalid formula") << "Invalid WP cut formula = '" << cut_str << "'.";
+      }
+    }
+    double operator()(const reco::BaseTau& tau, bool isPFTau) const {
+      if (!fn_) {
+        return value_;
+      }
+
+      if (isPFTau)
+        fn_->SetParameter(0, dynamic_cast<const reco::PFTau&>(tau).decayMode());
+      else
+        fn_->SetParameter(0, dynamic_cast<const pat::Tau&>(tau).decayMode());
+      fn_->SetParameter(1, tau.pt());
+      fn_->SetParameter(2, tau.eta());
+      return fn_->Eval(0);
+    }
+
+  private:
+    std::unique_ptr<TF1> fn_;
+    double value_;
+  };
+
+  class DeepTauCache {
+  public:
+    using GraphPtr = std::shared_ptr<tensorflow::GraphDef>;
+
+    DeepTauCache(const std::map<std::string, std::string>& graph_names, bool mem_mapped) {
+      for (const auto& graph_entry : graph_names) {
+        // Backend : to be parametrized from the python config
+        tensorflow::Options options{tensorflow::Backend::cpu};
+
+        const std::string& entry_name = graph_entry.first;
+        const std::string& graph_file = graph_entry.second;
+        if (mem_mapped) {
+          memmappedEnv_[entry_name] = std::make_unique<tensorflow::MemmappedEnv>(tensorflow::Env::Default());
+          const tensorflow::Status mmap_status = memmappedEnv_.at(entry_name)->InitializeFromFile(graph_file);
+          if (!mmap_status.ok()) {
+            throw cms::Exception("DeepTauCache: unable to initalize memmapped environment for ")
+                << graph_file << ". \n"
+                << mmap_status.ToString();
+          }
+
+          graphs_[entry_name] = std::make_unique<tensorflow::GraphDef>();
+          const tensorflow::Status load_graph_status =
+              ReadBinaryProto(memmappedEnv_.at(entry_name).get(),
+                              tensorflow::MemmappedFileSystem::kMemmappedPackageDefaultGraphDef,
+                              graphs_.at(entry_name).get());
+          if (!load_graph_status.ok())
+            throw cms::Exception("DeepTauCache: unable to load graph from ") << graph_file << ". \n"
+                                                                             << load_graph_status.ToString();
+
+          options.getSessionOptions().config.mutable_graph_options()->mutable_optimizer_options()->set_opt_level(
+              ::tensorflow::OptimizerOptions::L0);
+          options.getSessionOptions().env = memmappedEnv_.at(entry_name).get();
+
+          sessions_[entry_name] = tensorflow::createSession(graphs_.at(entry_name).get(), options);
+
+        } else {
+          graphs_[entry_name].reset(tensorflow::loadGraphDef(graph_file));
+          sessions_[entry_name] = tensorflow::createSession(graphs_.at(entry_name).get(), options);
+        }
+      }
+    };
+    ~DeepTauCache() {
+      for (auto& session_entry : sessions_)
+        tensorflow::closeSession(session_entry.second);
+    }
+
+    // A Session allows concurrent calls to Run(), though a Session must
+    // be created / extended by a single thread.
+    tensorflow::Session& getSession(const std::string& name = "") const { return *sessions_.at(name); }
+    const tensorflow::GraphDef& getGraph(const std::string& name = "") const { return *graphs_.at(name); }
+
+  private:
+    std::map<std::string, GraphPtr> graphs_;
+    std::map<std::string, tensorflow::Session*> sessions_;
+    std::map<std::string, std::unique_ptr<tensorflow::MemmappedEnv>> memmappedEnv_;
+  };
+
   constexpr int NumberOfOutputs = 4;
-}
+}  // namespace deep_tau
 
 namespace {
 
@@ -312,7 +448,7 @@ namespace {
     const edm::AssociationVector<reco::PFTauRefProd, std::vector<reco::PFTauTransverseImpactParameterRef>>*
         pfTauTransverseImpactParameters;
 
-    using BasicDiscr = deep_tau::DeepTauBase::BasicDiscriminator;
+    using BasicDiscr = deep_tau::BasicDiscriminator;
     std::map<BasicDiscr, size_t> indexMap;
     std::map<BasicDiscr, size_t> indexMapdR03;
 
@@ -805,27 +941,59 @@ namespace {
   };
 }  // anonymous namespace
 
-using bd = deep_tau::DeepTauBase::BasicDiscriminator;
-const std::map<bd, std::string> deep_tau::DeepTauBase::stringFromDiscriminator_{
-    {bd::ChargedIsoPtSum, "ChargedIsoPtSum"},
-    {bd::NeutralIsoPtSum, "NeutralIsoPtSum"},
-    {bd::NeutralIsoPtSumWeight, "NeutralIsoPtSumWeight"},
-    {bd::FootprintCorrection, "TauFootprintCorrection"},
-    {bd::PhotonPtSumOutsideSignalCone, "PhotonPtSumOutsideSignalCone"},
-    {bd::PUcorrPtSum, "PUcorrPtSum"}};
-const std::vector<bd> deep_tau::DeepTauBase::requiredBasicDiscriminators_ = {bd::ChargedIsoPtSum,
-                                                                             bd::NeutralIsoPtSum,
-                                                                             bd::NeutralIsoPtSumWeight,
-                                                                             bd::PhotonPtSumOutsideSignalCone,
-                                                                             bd::PUcorrPtSum};
-const std::vector<bd> deep_tau::DeepTauBase::requiredBasicDiscriminatorsdR03_ = {bd::ChargedIsoPtSum,
-                                                                                 bd::NeutralIsoPtSum,
-                                                                                 bd::NeutralIsoPtSumWeight,
-                                                                                 bd::PhotonPtSumOutsideSignalCone,
-                                                                                 bd::FootprintCorrection};
-
-class DeepTauId : public deep_tau::DeepTauBase {
+class DeepTauId : public edm::stream::EDProducer<edm::GlobalCache<deep_tau::DeepTauCache>> {
 public:
+  using TauDiscriminator = reco::TauDiscriminatorContainer;
+  using TauCollection = edm::View<reco::BaseTau>;
+  using CandidateCollection = edm::View<reco::Candidate>;
+  using TauRef = edm::Ref<TauCollection>;
+  using TauRefProd = edm::RefProd<TauCollection>;
+  using ElectronCollection = pat::ElectronCollection;
+  using MuonCollection = pat::MuonCollection;
+  using LorentzVectorXYZ = ROOT::Math::LorentzVector<ROOT::Math::PxPyPzE4D<double>>;
+  using Cutter = deep_tau::TauWPThreshold;
+  using CutterPtr = std::unique_ptr<Cutter>;
+  using WPList = std::vector<CutterPtr>;
+
+  struct Output {
+    std::vector<size_t> num_, den_;
+
+    Output(const std::vector<size_t>& num, const std::vector<size_t>& den) : num_(num), den_(den) {}
+
+    std::unique_ptr<TauDiscriminator> get_value(const edm::Handle<TauCollection>& taus,
+                                                const tensorflow::Tensor& pred,
+                                                const WPList* working_points,
+                                                bool is_online) const {
+      std::vector<reco::SingleTauDiscriminatorContainer> outputbuffer(taus->size());
+
+      for (size_t tau_index = 0; tau_index < taus->size(); ++tau_index) {
+        float x = 0;
+        for (size_t num_elem : num_)
+          x += pred.matrix<float>()(tau_index, num_elem);
+        if (x != 0 && !den_.empty()) {
+          float den_val = 0;
+          for (size_t den_elem : den_)
+            den_val += pred.matrix<float>()(tau_index, den_elem);
+          x = den_val != 0 ? x / den_val : std::numeric_limits<float>::max();
+        }
+        outputbuffer[tau_index].rawValues.push_back(x);
+        if (working_points) {
+          for (const auto& wp : *working_points) {
+            const bool pass = x > (*wp)(taus->at(tau_index), is_online);
+            outputbuffer[tau_index].workingPoints.push_back(pass);
+          }
+        }
+      }
+      std::unique_ptr<TauDiscriminator> output = std::make_unique<TauDiscriminator>();
+      reco::TauDiscriminatorContainer::Filler filler(*output);
+      filler.insert(taus, outputbuffer.begin(), outputbuffer.end());
+      filler.fill();
+      return output;
+    }
+  };
+
+  using OutputCollection = std::map<std::string, Output>;
+
   static constexpr float default_value = -999.;
 
   static const OutputCollection& GetOutputs() {
@@ -837,6 +1005,8 @@ public:
     };
     return outputs_;
   }
+
+  using BasicDiscriminator = deep_tau::BasicDiscriminator;
 
   const std::map<BasicDiscriminator, size_t> matchDiscriminatorIndices(
       edm::Event& event,
@@ -916,7 +1086,12 @@ public:
 
 public:
   explicit DeepTauId(const edm::ParameterSet& cfg, const deep_tau::DeepTauCache* cache)
-      : DeepTauBase(cfg, GetOutputs(), cache),
+      : tausToken_(consumes<TauCollection>(cfg.getParameter<edm::InputTag>("taus"))),
+        pfcandToken_(consumes<CandidateCollection>(cfg.getParameter<edm::InputTag>("pfcands"))),
+        vtxToken_(consumes<reco::VertexCollection>(cfg.getParameter<edm::InputTag>("vertices"))),
+        is_online_(cfg.getParameter<bool>("is_online")),
+        outputs_(GetOutputs()),
+        cache_(cache),
         electrons_token_(consumes<std::vector<pat::Electron>>(cfg.getParameter<edm::InputTag>("electrons"))),
         muons_token_(consumes<std::vector<pat::Muon>>(cfg.getParameter<edm::InputTag>("muons"))),
         rho_token_(consumes<double>(cfg.getParameter<edm::InputTag>("rho"))),
@@ -937,6 +1112,55 @@ public:
         save_inputs_(cfg.getParameter<bool>("save_inputs")),
         json_file_(nullptr),
         file_counter_(0) {
+    for (const auto& output_desc : outputs_) {
+      produces<TauDiscriminator>(output_desc.first);
+      const auto& cut_list = cfg.getParameter<std::vector<std::string>>(output_desc.first + "WP");
+      for (const std::string& cut_str : cut_list) {
+        workingPoints_[output_desc.first].push_back(std::make_unique<Cutter>(cut_str));
+      }
+    }
+
+    // prediscriminant operator
+    // require the tau to pass the following prediscriminants
+    const edm::ParameterSet& prediscriminantConfig = cfg.getParameter<edm::ParameterSet>("Prediscriminants");
+
+    // determine boolean operator used on the prediscriminants
+    std::string pdBoolOperator = prediscriminantConfig.getParameter<std::string>("BooleanOperator");
+    // convert string to lowercase
+    transform(pdBoolOperator.begin(), pdBoolOperator.end(), pdBoolOperator.begin(), ::tolower);
+
+    if (pdBoolOperator == "and") {
+      andPrediscriminants_ = 0x1;  //use chars instead of bools so we can do a bitwise trick later
+    } else if (pdBoolOperator == "or") {
+      andPrediscriminants_ = 0x0;
+    } else {
+      throw cms::Exception("TauDiscriminationProducerBase")
+          << "PrediscriminantBooleanOperator defined incorrectly, options are: AND,OR";
+    }
+
+    // get the list of prediscriminants
+    std::vector<std::string> prediscriminantsNames =
+        prediscriminantConfig.getParameterNamesForType<edm::ParameterSet>();
+
+    for (auto const& iDisc : prediscriminantsNames) {
+      const edm::ParameterSet& iPredisc = prediscriminantConfig.getParameter<edm::ParameterSet>(iDisc);
+      const edm::InputTag& label = iPredisc.getParameter<edm::InputTag>("Producer");
+      double cut = iPredisc.getParameter<double>("cut");
+
+      if (is_online_) {
+        TauDiscInfo<reco::PFTauDiscriminator> thisDiscriminator;
+        thisDiscriminator.label = label;
+        thisDiscriminator.cut = cut;
+        thisDiscriminator.disc_token = consumes<reco::PFTauDiscriminator>(label);
+        recoPrediscriminants_.push_back(thisDiscriminator);
+      } else {
+        TauDiscInfo<pat::PATTauDiscriminator> thisDiscriminator;
+        thisDiscriminator.label = label;
+        thisDiscriminator.cut = cut;
+        thisDiscriminator.disc_token = consumes<pat::PATTauDiscriminator>(label);
+        patPrediscriminants_.push_back(thisDiscriminator);
+      }
+    }
     if (version_ == 2) {
       using namespace dnn_inputs_v2;
       namespace sc = deep_tau::Scaling;
@@ -1015,14 +1239,88 @@ public:
     }
   }
 
-  static std::unique_ptr<deep_tau::DeepTauCache> initializeGlobalCache(const edm::ParameterSet& cfg) {
-    return DeepTauBase::initializeGlobalCache(cfg);
+  void produce(edm::Event& event, const edm::EventSetup& es) override {
+    edm::Handle<TauCollection> taus;
+    event.getByToken(tausToken_, taus);
+    edm::ProductID tauProductID = taus.id();
+
+    // load prediscriminators
+    size_t nPrediscriminants =
+        patPrediscriminants_.empty() ? recoPrediscriminants_.size() : patPrediscriminants_.size();
+    for (size_t iDisc = 0; iDisc < nPrediscriminants; ++iDisc) {
+      edm::ProductID discKeyId;
+      if (is_online_) {
+        recoPrediscriminants_[iDisc].fill(event);
+        discKeyId = recoPrediscriminants_[iDisc].handle->keyProduct().id();
+      } else {
+        patPrediscriminants_[iDisc].fill(event);
+        discKeyId = patPrediscriminants_[iDisc].handle->keyProduct().id();
+      }
+
+      // Check to make sure the product is correct for the discriminator.
+      // If not, throw a more informative exception.
+      if (tauProductID != discKeyId) {
+        throw cms::Exception("MisconfiguredPrediscriminant")
+            << "The tau collection has product ID: " << tauProductID
+            << " but the pre-discriminator is keyed with product ID: " << discKeyId << std::endl;
+      }
+    }
+
+    const tensorflow::Tensor& pred = getPredictions(event, taus);
+    createOutputs(event, pred, taus);
   }
 
-  static void globalEndJob(const deep_tau::DeepTauCache* cache_) { return DeepTauBase::globalEndJob(cache_); }
+  static std::unique_ptr<deep_tau::DeepTauCache> initializeGlobalCache(const edm::ParameterSet& cfg) {
+    const auto graph_name_vector = cfg.getParameter<std::vector<std::string>>("graph_file");
+    std::map<std::string, std::string> graph_names;
+    for (const auto& entry : graph_name_vector) {
+      const size_t sep_pos = entry.find(':');
+      std::string entry_name, graph_file;
+      if (sep_pos != std::string::npos) {
+        entry_name = entry.substr(0, sep_pos);
+        graph_file = entry.substr(sep_pos + 1);
+      } else {
+        entry_name = "";
+        graph_file = entry;
+      }
+      graph_file = edm::FileInPath(graph_file).fullPath();
+      if (graph_names.count(entry_name))
+        throw cms::Exception("DeepTauCache") << "Duplicated graph entries";
+      graph_names[entry_name] = graph_file;
+    }
+    bool mem_mapped = cfg.getParameter<bool>("mem_mapped");
+    return std::make_unique<deep_tau::DeepTauCache>(graph_names, mem_mapped);
+  }
+
+  static void globalEndJob(const deep_tau::DeepTauCache* cache_) {}
+
+  template <typename ConsumeType>
+  struct TauDiscInfo {
+    edm::InputTag label;
+    edm::Handle<ConsumeType> handle;
+    edm::EDGetTokenT<ConsumeType> disc_token;
+    double cut;
+    void fill(const edm::Event& evt) { evt.getByToken(disc_token, handle); }
+  };
+
+  // select boolean operation on prediscriminants (and = 0x01, or = 0x00)
+  uint8_t andPrediscriminants_;
+  std::vector<TauDiscInfo<pat::PATTauDiscriminator>> patPrediscriminants_;
+  std::vector<TauDiscInfo<reco::PFTauDiscriminator>> recoPrediscriminants_;
 
 private:
   static constexpr float pi = M_PI;
+
+  void createOutputs(edm::Event& event, const tensorflow::Tensor& pred, edm::Handle<TauCollection> taus) {
+    for (const auto& output_desc : outputs_) {
+      const WPList* working_points = nullptr;
+      if (workingPoints_.find(output_desc.first) != workingPoints_.end()) {
+        working_points = &workingPoints_.at(output_desc.first);
+      }
+      auto result = output_desc.second.get_value(taus, pred, working_points, is_online_);
+      event.put(std::move(result), output_desc.first);
+    }
+  }
 
   template <typename T>
   static float getValue(T value) {
@@ -1188,7 +1486,7 @@ private:
   }
 
 private:
-  tensorflow::Tensor getPredictions(edm::Event& event, edm::Handle<TauCollection> taus) override {
+  tensorflow::Tensor getPredictions(edm::Event& event, edm::Handle<TauCollection> taus) {
     // Empty dummy vectors
     const std::vector<pat::Electron> electron_collection_default;
     const std::vector<pat::Muon> muon_collection_default;
@@ -2429,6 +2727,33 @@ private:
   }
 
 private:
+  edm::EDGetTokenT<TauCollection> tausToken_;
+  edm::EDGetTokenT<CandidateCollection> pfcandToken_;
+  edm::EDGetTokenT<reco::VertexCollection> vtxToken_;
+  std::map<std::string, WPList> workingPoints_;
+  const bool is_online_;
+  OutputCollection outputs_;
+  const deep_tau::DeepTauCache* cache_;
+
+  const std::map<BasicDiscriminator, std::string> stringFromDiscriminator_{
+      {BasicDiscriminator::ChargedIsoPtSum, "ChargedIsoPtSum"},
+      {BasicDiscriminator::NeutralIsoPtSum, "NeutralIsoPtSum"},
+      {BasicDiscriminator::NeutralIsoPtSumWeight, "NeutralIsoPtSumWeight"},
+      {BasicDiscriminator::FootprintCorrection, "TauFootprintCorrection"},
+      {BasicDiscriminator::PhotonPtSumOutsideSignalCone, "PhotonPtSumOutsideSignalCone"},
+      {BasicDiscriminator::PUcorrPtSum, "PUcorrPtSum"}};
+  const std::vector<BasicDiscriminator> requiredBasicDiscriminators_{BasicDiscriminator::ChargedIsoPtSum,
+                                                                     BasicDiscriminator::NeutralIsoPtSum,
+                                                                     BasicDiscriminator::NeutralIsoPtSumWeight,
+                                                                     BasicDiscriminator::PhotonPtSumOutsideSignalCone,
+                                                                     BasicDiscriminator::PUcorrPtSum};
+  const std::vector<BasicDiscriminator> requiredBasicDiscriminatorsdR03_{
+      BasicDiscriminator::ChargedIsoPtSum,
+      BasicDiscriminator::NeutralIsoPtSum,
+      BasicDiscriminator::NeutralIsoPtSumWeight,
+      BasicDiscriminator::PhotonPtSumOutsideSignalCone,
+      BasicDiscriminator::FootprintCorrection};
+
   edm::EDGetTokenT<std::vector<pat::Electron>> electrons_token_;
   edm::EDGetTokenT<std::vector<pat::Muon>> muons_token_;
   edm::EDGetTokenT<double> rho_token_;
